feature: check giscus config
