+++ app/_components/Comment.tsx
@@ -1,29 +1,8 @@
-'use client';
-
 import config from 'public/config.json';
-import { useEffect, useRef } from 'react';
+import Giscus from './Giscus';
 
-function getScript() {
-  const giscusConfig = config.giscus;
-  const scriptElem = document.createElement('script');
-  const keys = Object.keys(giscusConfig);
-  const values = Object.values(giscusConfig);
-  for (let i = 0; i < keys.length; i += 1) {
-    scriptElem.setAttribute(keys[i], values[i].toString());
+export default function Comment() {
+  if (config.giscus) {
+    return <Giscus />;
   }
-
-  return scriptElem;
-}
-
-const giscusScript = getScript();
-
-export default function Giscus() {
-  const ref = useRef<HTMLDivElement>(null);
-
-  useEffect(() => {
-    if (!ref.current || ref.current.hasChildNodes()) return;
-    ref.current.appendChild(giscusScript);
-  }, []);
-
-  return <section ref={ref} />;
 }
